fix: add posthog error monitor
