+++ apps/remix/app/entry.client.tsx
@@ -1,21 +1,30 @@
-import { StrictMode, startTransition } from 'react';
+import { StrictMode, startTransition, useEffect } from 'react';
 
 import { i18n } from '@lingui/core';
 import { detect, fromHtmlTag } from '@lingui/detect-locale';
 import { I18nProvider } from '@lingui/react';
+import posthog from 'posthog-js';
 import { hydrateRoot } from 'react-dom/client';
 import { HydratedRouter } from 'react-router/dom';
-import { Theme, ThemeProvider } from 'remix-themes';
-import { match } from 'ts-pattern';
 
+import { extractPostHogConfig } from '@documenso/lib/constants/feature-flags';
 import { dynamicActivate } from '@documenso/lib/utils/i18n';
 
-async function main() {
-  const theme = match(document.documentElement.getAttribute('data-theme'))
-    .with('dark', () => Theme.DARK)
-    .with('light', () => Theme.LIGHT)
-    .otherwise(() => null);
+function PosthogInit() {
+  const postHogConfig = extractPostHogConfig();
 
+  useEffect(() => {
+    if (postHogConfig) {
+      posthog.init(postHogConfig.key, {
+        api_host: postHogConfig.host,
+      });
+    }
+  }, []);
+
+  return null;
+}
+
+async function main() {
   const locale = detect(fromHtmlTag('lang')) || 'en';
 
   await dynamicActivate(locale);
@@ -25,10 +34,10 @@ async function main() {
       document,
       <StrictMode>
         <I18nProvider i18n={i18n}>
-          <ThemeProvider specifiedTheme={theme} themeAction="/api/theme">
-            <HydratedRouter />
-          </ThemeProvider>
+          <HydratedRouter />
         </I18nProvider>
+
+        <PosthogInit />
       </StrictMode>,
     );
   });
