+++ apps/remix/app/entry.server.tsx
@@ -7,13 +7,11 @@ import type { RenderToPipeableStreamOptions } from 'react-dom/server';
 import { renderToPipeableStream } from 'react-dom/server';
 import type { AppLoadContext, EntryContext } from 'react-router';
 import { ServerRouter } from 'react-router';
-import { ThemeProvider } from 'remix-themes';
 
 import { APP_I18N_OPTIONS } from '@documenso/lib/constants/i18n';
 import { dynamicActivate, extractLocaleData } from '@documenso/lib/utils/i18n';
 
 import { langCookie } from './storage/lang-cookie.server';
-import { themeSessionResolver } from './storage/theme-session.server';
 
 export const streamTimeout = 5_000;
 
@@ -32,10 +30,6 @@ export default async function handleRequest(
 
   await dynamicActivate(language);
 
-  const { getTheme } = await themeSessionResolver(request);
-
-  const theme = getTheme();
-
   return new Promise((resolve, reject) => {
     let shellRendered = false;
     const userAgent = request.headers.get('user-agent');
@@ -47,9 +41,7 @@ export default async function handleRequest(
 
     const { pipe, abort } = renderToPipeableStream(
       <I18nProvider i18n={i18n}>
-        <ThemeProvider specifiedTheme={theme} themeAction="/api/theme">
-          <ServerRouter context={routerContext} url={request.url} />
-        </ThemeProvider>
+        <ServerRouter context={routerContext} url={request.url} />
       </I18nProvider>,
       {
         [readyOption]() {
