+++ apps/remix/package.json
@@ -48,8 +48,8 @@
     "luxon": "^3.4.0",
     "papaparse": "^5.4.1",
     "plausible-tracker": "^0.3.9",
-    "posthog-js": "^1.75.3",
-    "posthog-node": "^3.1.1",
+    "posthog-js": "^1.223.3",
+    "posthog-node": "^4.7.0",
     "react": "^18",
     "react-call": "^1.3.0",
     "react-dom": "^18",
