+++ packages/lib/universal/upload/get-file.server.ts
@@ -2,6 +2,8 @@ import { DocumentDataType } from '@prisma/client';
 import { base64 } from '@scure/base';
 import { match } from 'ts-pattern';
 
+import { getPresignGetUrl } from './server-actions';
+
 export type GetFileOptions = {
   type: DocumentDataType;
   data: string;
@@ -30,8 +32,6 @@ const getFileFromBytes64 = (data: string) => {
 };
 
 const getFileFromS3 = async (key: string) => {
-  const { getPresignGetUrl } = await import('./server-actions');
-
   const { url } = await getPresignGetUrl(key);
 
   const response = await fetch(url, {
