+++ packages/lib/universal/upload/update-file.ts
@@ -2,6 +2,8 @@ import { DocumentDataType } from '@prisma/client';
 import { base64 } from '@scure/base';
 import { match } from 'ts-pattern';
 
+import { getAbsolutePresignPostUrl } from './server-actions';
+
 export type UpdateFileOptions = {
   type: DocumentDataType;
   oldData: string;
@@ -37,8 +39,6 @@ const updateFileWithBytes64 = (data: string) => {
 };
 
 const updateFileWithS3 = async (key: string, data: string) => {
-  const { getAbsolutePresignPostUrl } = await import('./server-actions');
-
   const { url } = await getAbsolutePresignPostUrl(key);
 
   const response = await fetch(url, {
